[#74] feat(auth): web용 간편 로그인 방식 파일 분리
